pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-security-mcp"
version = "0.1.0"
description = "Smart Security and Surveillance Assistant MCP Server"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
keywords = ["mcp", "smart-home", "security", "esp32", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "pylint>=2.17.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/smart-security-mcp"
Documentation = "https://github.com/yourusername/smart-security-mcp/blob/main/README.md"
Repository = "https://github.com/yourusername/smart-security-mcp.git"
Issues = "https://github.com/yourusername/smart-security-mcp/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pylint.messages_control]
disable = [
    "C0111",  # Missing module docstring
    "C0103",  # Invalid name
]
